22-5-23

+++ app/Http/Controllers/InvoiceController.php
@@ -10,6 +10,7 @@ use App\Models\InvoicesAttachments;
 use App\Models\Section;
 use Illuminate\Http\Request;
 use App\Traits\SessionMessages;
+use Spatie\Permission\Traits\HasRoles;
 
 class InvoiceController extends Controller
 {
@@ -19,8 +20,11 @@ class InvoiceController extends Controller
      */
     public function index()
     {
-        $data=Invoice::all();
-        return view('invoices.invoices',compact('data'));
+        if (auth()->user()->hasPermissionTo('show invoices')){
+            $data=Invoice::all();
+            return view('invoices.invoices',compact('data'));
+        }
+        return abort(404);
     }
 
     /**
@@ -28,8 +32,12 @@ class InvoiceController extends Controller
      */
     public function create()
     {
-        $sections=Section::whereHas('product')->get();
-        return view('invoices.add-invoice',compact('sections'));
+        if (auth()->user()->hasPermissionTo('add invoices')){
+
+            $sections=Section::whereHas('product')->get();
+            return view('invoices.add-invoice',compact('sections'));
+        }
+        return abort(404);
     }
 
     /**
@@ -37,19 +45,22 @@ class InvoiceController extends Controller
      */
     public function store(InvoiceRequest $request)
     {
-        $data=$request->validated();
-        $result=array_merge(['status'=>'3'],$data);
-        $lastID=Invoice::create($result);
-        $full=array_merge(['user'=>auth()->user()->name,'invoice_id'=>$lastID->id],$result);
-        $detect=InvoiceDetail::create($full);
-        if ($request->hasFile('pic')){
-            $photo=$request->pic->store('','invoices');
-            $fullData=array_merge(['file_name'=>$photo],$full);
-            InvoicesAttachments::create($fullData);
-        }
+        if (auth()->user()->hasPermissionTo('add invoices')){
+            $data=$request->validated();
+            $result=array_merge(['status'=>'3'],$data);
+            $lastID=Invoice::create($result);
+            $full=array_merge(['user'=>auth()->user()->name,'invoice_id'=>$lastID->id],$result);
+            $detect=InvoiceDetail::create($full);
+            if ($request->hasFile('pic')){
+                $photo=$request->pic->store('','invoices');
+                $fullData=array_merge(['file_name'=>$photo],$full);
+                InvoicesAttachments::create($fullData);
+            }
 
-        $this->messages($detect);
-        return redirect()->back();
+            $this->messages($detect);
+            return redirect()->back();
+        }
+        return abort(404);
     }
 
     /**
@@ -57,16 +68,19 @@ class InvoiceController extends Controller
      */
     public function show($id)
     {
-        $invoice=Invoice::findOrFail($id);
-        $invoiceDetails=InvoiceDetail::where('invoice_number',$invoice->invoice_number)->get();
-        $invoiceAttachment=InvoicesAttachments::where('invoice_number',$invoice->invoice_number)->get();
-        if ($invoiceDetails->isNotEmpty() && $invoiceAttachment->isNotEmpty()){
-            return view('invoices.show-invoice',['invoice'=>$invoice,'invoiceDetails'=>$invoiceDetails[0],'invoiceAttachments'=>$invoiceAttachment[0]]);
-        }elseif($invoiceDetails->isNotEmpty()){
-            return view('invoices.show-invoice',['invoice'=>$invoice,'invoiceDetails'=>$invoiceDetails[0]]);
-        }else{
-            return view('invoices.show-invoice',['invoice'=>$invoice]);
+        if (auth()->user()->hasPermissionTo('show invoices')){
+            $invoice=Invoice::findOrFail($id);
+            $invoiceDetails=InvoiceDetail::where('invoice_number',$invoice->invoice_number)->get();
+            $invoiceAttachment=InvoicesAttachments::where('invoice_number',$invoice->invoice_number)->get();
+            if ($invoiceDetails->isNotEmpty() && $invoiceAttachment->isNotEmpty()){
+                return view('invoices.show-invoice',['invoice'=>$invoice,'invoiceDetails'=>$invoiceDetails[0],'invoiceAttachments'=>$invoiceAttachment[0]]);
+            }elseif($invoiceDetails->isNotEmpty()){
+                return view('invoices.show-invoice',['invoice'=>$invoice,'invoiceDetails'=>$invoiceDetails[0]]);
+            }else{
+                return view('invoices.show-invoice',['invoice'=>$invoice]);
+            }
         }
+        return abort(404);
     }
 
     /**
@@ -74,9 +88,12 @@ class InvoiceController extends Controller
      */
     public function edit($id)
     {
-        $invoice=Invoice::findOrFail($id);
-        $sections=Section::whereHas('product')->get();
-        return view('invoices.edit-invoice',['invoice'=>$invoice,'sections'=>$sections]);
+        if (auth()->user()->hasPermissionTo('edit invoices')){
+            $invoice=Invoice::findOrFail($id);
+            $sections=Section::whereHas('product')->get();
+            return view('invoices.edit-invoice',['invoice'=>$invoice,'sections'=>$sections]);
+        }
+        return abort(404);
     }
 
     /**
@@ -84,17 +101,20 @@ class InvoiceController extends Controller
      */
     public function update(InvoiceEditRequest $request,$id)
     {
-        $data=$request->validated();
-        Invoice::where('id',$id)->update($data);
-        $detect=InvoiceDetail::where('invoice_number',$data)->update([
-            'invoice_number'=>$data['invoice_number'],
-            'product'=>$data['product'],
-            'section_id'=>$data['section_id'],
-            'status'=>$data['status'],
-            'note'=>$data['note'],
-        ]);
-        $this->messages($detect);
-        return redirect()->route('invoices.index');
+        if (auth()->user()->hasPermissionTo('edit invoices')){
+            $data=$request->validated();
+            Invoice::where('id',$id)->update($data);
+            $detect=InvoiceDetail::where('invoice_number',$data)->update([
+                'invoice_number'=>$data['invoice_number'],
+                'product'=>$data['product'],
+                'section_id'=>$data['section_id'],
+                'status'=>$data['status'],
+                'note'=>$data['note'],
+            ]);
+            $this->messages($detect);
+            return redirect()->route('invoices.index');
+        }
+        return abort(404);
     }
 
     /**
@@ -102,19 +122,22 @@ class InvoiceController extends Controller
      */
     public function destroy($id)
     {
-        $data=Invoice::findOrFail($id);
-        $data->delete();
-        $detail=InvoiceDetail::where('invoice_number',$data->invoice_number);
-        $attach=InvoicesAttachments::where('invoice_number',$data->invoice_number);
-        if ($detail){
-            $detail->delete();
-            if ($attach){
-                $fName=$attach->get();
-                unlink('storage/invoices/'.$fName[0]->file_name);
-                $attach->delete();
+        if (auth()->user()->hasPermissionTo('delete invoices')){
+            $data=Invoice::findOrFail($id);
+            $data->delete();
+            $detail=InvoiceDetail::where('invoice_number',$data->invoice_number);
+            $attach=InvoicesAttachments::where('invoice_number',$data->invoice_number);
+            if ($detail){
+                $detail->delete();
+                if ($attach){
+                    $fName=$attach->get();
+                    unlink('storage/invoices/'.$fName[0]->file_name);
+                    $attach->delete();
+                }
             }
+            $this->messages($data);
+            return redirect()->back();
         }
-        $this->messages($data);
-        return redirect()->back();
+        return abort(404);
     }
 }
